+++ Test/bootstrap.php
@@ -129,3 +129,26 @@ namespace Magento\Catalog\Model\ResourceModel\Product {
         }
     }
 }
+
+namespace Magento\Eav\Model\ResourceModel\Entity\Attribute {
+    if (!class_exists(CollectionFactory::class)) {
+        class CollectionFactory
+        {
+            protected $_objectManager;
+            protected $_instanceName;
+
+            public function __construct(
+                \Magento\Framework\ObjectManagerInterface $objectManager,
+                $instanceName = Collection::class
+            ) {
+                $this->_objectManager = $objectManager;
+                $this->_instanceName = $instanceName;
+            }
+
+            public function create(array $data = [])
+            {
+                return $this->_objectManager->create($this->_instanceName, $data);
+            }
+        }
+    }
+}